+++ server/src/modules/queues/constants.js
@@ -1,0 +1,11 @@
+const QUEUE_EVENTS = {
+  VIDEO_UPLOADED: "video.uploaded",
+  VIDEO_PROCESSING: "video.processing",
+  VIDEO_PROCESSED: "video.processed",
+  VIDEO_HLS_CONVERTING: "video.hls-converting",
+  VIDEO_HLS_CONVERTED: "video.hls.converted",
+  VIDEO_WATERMARKING: "video.watermarking",
+  VIDEO_WATERMARKED: "video.watermarked",
+};
+
+module.exports = { QUEUE_EVENTS };
